change password action added in user controller

+++ Personal_Expense_Tracking_System_Web_Api/Models/Categories.cs
@@ -5,7 +5,7 @@ namespace Personal_Expense_Tracking_System_Web_Api.Models
     public class Categories
     {
         [Key]
-        public int CategoryID { get; set; }
+        public long CategoryID { get; set; }
         public string CategoryName { get; set; }
     }
 }
