+++ Personal_Expense_Tracking_System_Web_Api/Models/ExpenseCategories.cs
@@ -5,7 +5,7 @@ namespace Personal_Expense_Tracking_System_Web_Api.Models
     public class ExpenseCategories
     {
         [Key]
-        public int ExpenseCategoryID { get; set; }
+        public long ExpenseCategoryID { get; set; }
         public string CategoryName { get; set; }
     }
 }
